# Fix cursor history loss on SeqDecoder.