Fixes #2283 

This PR fixes the bug introduced in #1719 on SeqDecoder cursor history. 

I tried to keep the performance optimization recently introduced. Another possibility is to revert the implementation to the original code which looks less hacky in my opinion.


```
#before
[info] Benchmark                                      Mode  Cnt      Score      Error  Units
[info] NumberParsingBenchmark.decodeManyBigDecimals  thrpt    5   1459.271 ±  371.206  ops/s
[info] NumberParsingBenchmark.decodeManyBigInts      thrpt    5   1666.720 ±  228.149  ops/s
[info] NumberParsingBenchmark.decodeManyDoubles      thrpt    5  10748.210 ±  320.140  ops/s
[info] NumberParsingBenchmark.decodeManyLongs        thrpt    5  11131.330 ± 1333.619  ops/s

#after   
[info] Benchmark                                      Mode  Cnt      Score      Error  Units
[info] NumberParsingBenchmark.decodeManyBigDecimals  thrpt    5   1503.769 ±  115.695  ops/s
[info] NumberParsingBenchmark.decodeManyBigInts      thrpt    5   1695.076 ±  449.295  ops/s
[info] NumberParsingBenchmark.decodeManyDoubles      thrpt    5  10075.492 ± 1772.034  ops/s
[info] NumberParsingBenchmark.decodeManyLongs        thrpt    5  12766.322 ±  994.195  ops/s
```